test: 결제 실패 테스트 추가
