+++ src/test/java/roomescape/integration/IntegrationTest.java
@@ -12,9 +12,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
-import roomescape.config.TestConfig;
 import roomescape.helper.CookieProvider;
 import roomescape.helper.DatabaseCleaner;
 import roomescape.helper.domain.MemberFixture;
@@ -25,13 +23,12 @@ import roomescape.helper.domain.WaitingFixture;
 import roomescape.service.payment.PaymentClient;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@Import(TestConfig.class)
 @ActiveProfiles("test")
 public abstract class IntegrationTest {
     @LocalServerPort
     int port;
 
-    @Autowired
+    @MockBean
     protected PaymentClient paymentClient;
 
     @Autowired
